update-scaffolding+add-jwt-required-to-all-api
