+++ server/helpers.py
@@ -1,0 +1,17 @@
+import datetime
+import json
+
+from bson import ObjectId
+
+
+class JSONEncoder(json.JSONEncoder):
+    """extend json-encoder class"""
+
+    def default(self, o):
+        if isinstance(o, ObjectId):
+            return str(o)
+        if isinstance(o, set):
+            return list(o)
+        if isinstance(o, datetime.datetime):
+            return str(o)
+        return json.JSONEncoder.default(self, o)
